+++ amqpstorm/tests/functional/exchange_tests.py
@@ -7,11 +7,10 @@ except ImportError:
 
 import amqpstorm
 from amqpstorm import Connection
+from amqpstorm.tests.functional import HOST
+from amqpstorm.tests.functional import USERNAME
+from amqpstorm.tests.functional import PASSWORD
 from amqpstorm.tests.utility import MockLoggingHandler
-
-HOST = '127.0.0.1'
-USERNAME = 'guest'
-PASSWORD = 'guest'
 
 logging.basicConfig(level=logging.DEBUG)